Fixed an issue in generateImagePair_ND_mc.m where the images could become saturated for large particles. The fix was to invoke the matlab command mat2gray to rescale the intensities in each image from 0 - 1, and then multiply by intmax('uint16').

+++ generateImagePair_ND_mc.m
@@ -197,11 +197,11 @@ end
 
 
 % Convert the first image to 16 bit and save it to the output variable.
-IMAGE_01 = uint16( (2^16 - 1) .* ...
+IMAGE_01 = ( (2^16 - 1) .* ...
     Image1Cropped .* 2.8 ^ 2 / PARTICLE_DIAMETER_MEAN ^ 2);
 
 % Convert the second image to 16 bit and save it to the output variable.
-IMAGE_02 = uint16( (2^16 - 1) .* Image2Cropped .* ...
+IMAGE_02 = ( (2^16 - 1) .* Image2Cropped .* ...
     2.8 ^ 2 ./ PARTICLE_DIAMETER_MEAN ^ 2);
